harvester/requirements.txt
# requirements.txt
python-nmap
tk
requests
